Fix missing hosts file entry for development
In development, running VMs, if a request is made to the QA app,
it calls the api with the same hostname as the QA app.  If this is
requested on "local.dp.la" then the internal request back to itself
will fail if that hostname is not in /etc/hosts on the server.