Dynamically set WordPress WP_HOME & WP_SITEURL
This should allow us to temporarily use another CNAME to verify
the WordPress install before we switch it to prod. The navigation
links should now also work as expected. Once we get closer to
the final production deployment this should be switched back to
use the frontend_hostname variable.

Also fixes an issue where there's an extra slash in the navigation
for portal and exhibition links in WordPress.