Add Ingestion 2
- Change dev environment for ingestion2
- Add Redis role
- Add Tomcat common role
- Add Solr role
- Add Marmotta role
- Add ingestion_app role

Various notes:

Tomcat 7 is used because Marmotta requires it.
Solr is installed from a tarfile, instead of an Ubuntu package,
because Ubuntu's package relies on the tomcat6 package.

Ubuntu 14.04 LTS (trusty) is specified for new Ingestion 2 servers.
The only reason we're still using 12.04 LTS for everything else
is because of BigCouch, and that's going to go away, so let's look
to the future and use 14.04 where we can, now.  Let's start now
and get everything to work with 14.04 for new work.

API authentication is disabled in the development VMs for Ingestion 2,
because we're not running BigCouch in the development VMs, in order
to save memory.

Please note that the new 2-VM Vagrant configuration will go away after
Ingestion 2 hits production and we can go back to our usual
loadbalanced development VMs.

PostgreSQL has been tuned to serve as Marmotta's database.  Values that
have become more important with the addition of Ingestion 2 have been
parameterized.

There is a new README file, README-ingestion2.md, that is linked from
the main README.  This covers how to set up your development VMs for
Ingestion 2.

When Ingestion 2 hits production, we will go back to the usual VM
configuration with the loadbalancer, and remove things we no longer
need, like BigCouch and the "contentqa" environment.
README-ingestion2.md will be incorporated into the main README and
deleted.

Our near-term TODO list includes some further Marmotta tuning and
dedicated playbooks for Heidrun deployment, but Marmotta seems to run
well enough for now, and deployment can be done easily enough with the
existing ingestion_app_* playbooks.