Fix incorrect rsync options for deployments
Rsync is used in various deployment scripts to copy files between
source, build, and target directories.

When rsync is called with `--checksum', the `u' option should
not be passed, because this overrides the desired checksum
behavior.

This was causing  files modified on the server more recently than their
sources in the local development working copy to fail to be overwritten.
In at least one case, we modify a Gemfile in the server's build
directory after it's copied there, which results in its timestamp being
newer than that of the source.  The next time it's deployed, it will
fail to be copied if the `u' switch is given to rsync.

Here is an example of rsync's behavior with and without `u', combined
with `checksum':

  $ mkdir a b
  $ echo line1 > a/file
  $ rsync -a a/ b
  $ cat b/file
  line1
  $ echo line2 >> b/file
  $ rsync -ruptv --checksum a/ b
  # b/file is _not_ updated to match a/file ...
  $ cat b/file
  line1
  line2
  $ rsync -rptv --checksum a/ b
  # b/file _is_ updated to match a/file ...
  $ cat b/file
  line1

This change has been tested with local deployments of api, frontend,
ingestion_app, and wordpress, by modifying files on the target host
and observing that they get overwritten.