Change solr configuration in ingestion_app role
This change is a companion to the changes in
https://github.com/dpla/KriKri/pull/103

The Solr configuration for IndexService has been parameterized, so add
a Solr section to settings.local.yml in heidrun's ./config directory.
Stop managing the state of solr.yml in that directory.  Note the
patching of the Blacklight configuration in the aforementioned pull
request.

With these changes you should observe the following in the rails
console in development:
    [1] pry(main)> Krikri::Settings.solr.url
    => "http://dev2:8080/solr"
    [2] pry(main)> Blacklight.solr_config
    => {:url=>"http://dev2:8080/solr"}