Change Exhibitions cache strategy
Change proxy caching of the Exhibitions site to forego caching of HTML,
but keep caching static files.

Remove an `expires max' directive from the cms server's virtual host for
`omeka', to allow a browser "refresh" to cause the proxy to hit the CMS
with an if-modified-since request, achieving a nicer balance between
caching and refreshability.