Improve Solr index and configuration operations
- Delete Solr index when there is a schema change.
- Reload the Solr core instead of restarting Tomcat.

There is no standalone playbook for clearing the Solr index, which would
be nice, but which would take longer to put together than I can afford
right now, because I'd have to spend more effort to avoid duplicating
code in that case.

After clearing the index, you have to run a new indexing operation,
which is out of the scope of the `automation` project.  Just note that
that has to happen.  It seems that a first-time indexing run after the
schema is changed, the index is deleted, and the Solr core is restarted
requires a reload of the core again for the added documents to show up.
I don't have an explanation for that, and I'm not sure whether it has to
do with deleting the index, or something that should be done when the
records are re-indexed.