Show forwarded IP addresses of proxied http requests
In the site proxy, frontend, and API NGINX reverse proxies, log the
"real" IP addresses of the clients, instead of the IP addresses of the
loadbalancers or proxies in front of the NGINX servers.

This assists with troubleshooting, because it allows us to get a better
sense of the clients behind certain access patterns.