Add playbooks to configure frontend app
Add playbooks to provision and configure the operating environment and
services surrounding the frontend app.

A playbook already existed to deploy the frontend, but there was no
playbook dedicated to configuring Unicorn, NGINX, package dependencies,
and so on.

This playbook runs everything for the frontend role, so it will
configure as well as deploy it.