Add deployment roles, tasks, etc. for OEI site (fixes #7969)
This adds the deployment configuration for the Open Ebooks Initiative site.
The deployment pulls from a private GitHub repository for the source, checks
it out, and synchronizes the files with an S3 bucket. It is not designed to
be used in the development environment.

It also adds a localhost inventory and a playbook for deployments from your
local machine, which requires editing the associated local.yml file and
installation of s3cmd >= 1.5.0.