Change how 'cms' group is handled in pg_hba.conf (fixes #117)
In the template for PostgreSQL's pg_hba.conf, change how entries for the
'cms' inventory group are added.  Use "if <group> is defined" rather
than "if ingestion2 is not defined" in order to be consistent with the
other 'host' lines in the file.

This also has the potential to make the PostgreSQL configuration more
flexible with role-specific inventory files that we might want to create
in the future, stating the objective in the active voice.