Install private assets only if requested
Add variables that  governs whether the private dpla_frontend_assets gem
is included for branding-related assets.  This defuaults to false.

Addresses one of the problems raised in
https://github.com/dpla/automation/issues/117

When the gem is included, we write it into the Gemfile, which has been
modified elsewhere (in the app repositories) to not include the gem.

Also:
* Fail build scripts if bundle install fails
* Sneak in some documentation improvements about deployment ssh keys.