Allow setting of non-inventory database host
For various roles, add variables that allow for database hosts that
exist outside of the inventory.  A good example is an Amazon RDS
database instance that's not used in development.

The variables default to false so that normally the PostgreSQL database
host is used.

We also determine conditionally whether to put public-facing sites into
maintenance mode, based on whether they use the self-configured or
"alternate" database server.