Fix pkill task to stop old delayed_job process
`pkill -f' actually returns an error status if the process did not exist,
although it prints "Killed" to the terminal (not stdout or stderr -- I'm
not sure which file it goes to).

Interestingly, plain `pkill' (without -f) does _not_ print "Killed" if
the process did not exist, and it exits with 1.

The status that `pkill -f' returns if the process did not exist is
137, if a SIGKILL (-9) fails.  See http://tldp.org/LDP/abs/html/exitcodes.html

The ansible `command' module was giving the return code as 1 in this
case, instead of 137. The `shell' module, on the other hand, returns -9,
which is not what we'd expect (137) but is better than 1, which is
indistinguishable from other kinds of failures.