Add ImageMagick policy.xml file
Add ImageMagick file: /etc/ImageMagick/policy.xml

The immediate concern is to fix the recently-announced ImageTragick
exploit:
* https://imagetragick.com/
* https://bugs.launchpad.net/ubuntu/%2Bsource/imagemagick/%2Bbug/1578398

Note the conversation in the Ubuntu bugtracker in which the maintainer
critiques the response from the ImageMagick upstream developers. Even
though they have released a patch version (6.9.3-10) that they claim
fixes the problem, the maintainer advocates that it is not sufficient.
He provides some alternative approaches in
http://www.openwall.com/lists/oss-security/2016/05/03/19, including the
removal of the "delegate" functionality altogether, which does not seem
appropriate.

The fact that this debate is going on indicates to me that it is a good
idea for us to have our own policy.xml file whatever the case to be have
more control over ImageMagick's workings.