Specify explicit PostgreSQL version in roles
In roles that install the postgresql-client package, specify the package
name for the specific PostgreSQL version.  We use the apt.postgresql.org
repository, which gives you version 9.5 if you specify the apt
metapackage "postgresql-client".  We need version 9.4.

It would be nice to specify the pg_version variable in one place, but
there's no convenient place to do that outside of `vars/main.yml` in
each relevant role.