Stop clearing NGINX proxy cache unnecessarily
In v3.3.0, on 2016-05-31, the NGINX cache configuration was changed to
make it safer for the cache to hold on to assets for a longer time.
Pages that have to be fresh are no longer cached. We therefore can
remove the role tasks that used to clear the cache. We still clear the
Memcache cache where necessary, and we still have the playbook for doing
it manually.

Addresses https://issues.dp.la/issues/8157