Install geocoder config for ingestion
Install the geocoder.yml configuration file that is needed by the
Adumbla gem for enrichments in the ingestion application.

The installed configuration will not work in a development environment
until the developer sets the `ingestion_app_geo_host` variable to
something suitable for the development environment (for example, in the
`ansible/group_vars/development` file). The Twofishes geocoding server
is not installed into the development VMs and must be managed
separately.