Add new variable for all-in-one hosts
Add a new variable, `all_in_one_hosts`, which governs whether each host
in the inventory is self-contained, with no dependencies on other
servers. This defaults to False, so no changes should be necessary for
typical developer usage. This is for particular DPLA-specific use cases,
like QA, and is not relevant to either production or development VM
environments.

Also normalize some of the dict variable dereferencing syntax throughout
various files, for a cleaner look, and for consistency. For example,
change `groups['cms']` to `groups.cms`.