Fix mount / unmount in ec2_swapper role
Fix the usage of the `mount` module in the `ec2_swapper` role so that it
removes Ubuntu's `fstab` entry for `/mnt` when it unmounts it.

Also remove the unnecessary `file` task to create the `/swap` directory,
because the `mount` module will create it lower down in the file with
`state=mounted`.