Remove unwanted template task from marmotta role
Remove a `template` task that was causing errors due to its file being
missing in the `marmotta` role.

Management of a kernel parameters file was moved to the `tomcat_common`
role in commit caab9bef6a25 on 2016-09-21.  The `template` task for this
file was left in the `marmotta` role although it had been copied to
`tomcat_common`.  The template file had been moved and deleted from the
`marmotta` role.