Reintroduce check for ec2 /swap directory
Add back the task in the `ec2_swapper` role that checks that the `/swap`
directory exists. This was removed in 62631a563a30 with the assumption
that a later `mount` task would ensure its existence; overlooking the
fact that the `mount` task is not run on a t2 or m4 instance.