Improve jenkins role (#242)
Improve the jenkins role. This is for managing the state of a Jenkins master.

Add the s3cmd utility (a Python module) which is useful at least for ingestion3-related file transfers, if not for more things in the future.

Assuming we only have to worry about one version of Python relative to Jenkins's scripts, just set the pyenv python version globally.

Also change the global node version manager (nvm) installer script (used by other roles in automation) to write a per-user log file so that multiple per-user installs on one system are possible without file permissions errors.