Improve PSS deployment with local source (#250)
Improve the PSS deployment using the "lcoal source" option so that it's
easier to simulate the production environment with the development VMs.

Copying the configuration files from the developer's source code working
copy was confusing, because they're normally written from the .j2
templates. If you want to use "local source" and also have config files
that are similar to what you'd get in production, it's easier to manage
that by updating your group_vars/development file with the settings you
want.