Fix staging api wait_for task
In staging api deployment, use ansible_ssh_host because sometimes we
like to specify the hostname with that variable in inventory instead of
having to modify our /etc/hosts files. ansible_ssh_host should be
defined in either case, but inventory_host is not going to work if you
don't have a matching /etc/hosts entry.