Remove obsolete roles
Remove obsolete roles for services that are no longer used.

The api and dbnode roles are obsolete now with the rollout of the new
API application.

The elasticsearch_index role became obsolete when we converted to
Elasticsearch 6, and since we're now using sparkindexer.

The exporter role is obsolete because we refer people to our S3 dataset
and are working on an alternate exporter in Spark.